requirements.txt
click==7.1.2
pyyaml
Kivy==2.0.0
RPi.GPIO==0.7.0
pysdl2